Adding support for handling multiple transport protocols (#12967)
* Adding support for more than one protocol for transport


* Adding CHANGELOG entry


* Addressing comments


* Addressing comments


* Removing determineTransportProtocol


* Determine transport protocol only on first byte reference


* Making InboundBytesHandler closeable


* Fixing close() for InboundPipeline


* Adding DeprecatedAPI annotation to japicmp task


* Fixing for detect breaking changes workflow


* Fixing recursion


---------
